Add auth to workmode pages and remove oldAPI files
