+++ frontend/components/About.js
@@ -1,0 +1,7 @@
+export default function About() {
+  return (
+    <>
+      <p>Hey, I'm the about page now</p>
+    </>
+  );
+}
