+++ frontend/components/EasySort.js
@@ -1,23 +1,50 @@
 import SortableList, { SortableItem } from 'react-easy-sort';
 import styled from 'styled-components';
+import Link from 'next/link';
+import Button from './styles/Button';
 
 const SortableStlyes = styled.div`
   .list {
     display: grid;
-    grid-template-columns: auto auto;
+    grid-template-columns: auto auto auto auto;
     grid-template-rows: auto;
-  }
-
-  .item {
-    max-width: fit-content;
-    background: green;
-    user-select: none;
-    /* pointer-events: none; */
+    grid-gap: 20px;
+    align-items: center;
   }
 
   .dragged {
     background: red;
   }
+`;
+
+const SortGridItem = styled.div`
+  max-width: fit-content;
+  padding: 8px;
+  overflow: hidden;
+  border-radius: 30px;
+  border: 2px solid #00000066;
+  user-select: none;
+  position: relative;
+`;
+
+const NameWrapper = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  text-align: center;
+  background: hsl(0deg 0% 0% / 50%);
+  width: 100%;
+
+  p {
+    color: white;
+    margin: 0;
+    font-size: 0.8rem;
+  }
+`;
+
+const SortButtonContainer = styled.div`
+  display: flex;
+  flex-direction: column;
 `;
 
 export default function EasySort({ photos, onSortEnd, handleRemovePhoto }) {
@@ -31,16 +58,31 @@ export default function EasySort({ photos, onSortEnd, handleRemovePhoto }) {
       >
         {photos.map((photo, idx) => (
           <SortableItem key={photo?.id}>
-            <div className="item">
+            <SortGridItem>
               <img
                 src={photo?.image?.publicUrlTransformed}
                 alt={`${photo?.name} Thumb`}
               />
-              <p>{photo?.name}</p>
-              <button type="button" data-idx={idx} onClick={handleRemovePhoto}>
-                Remove this photo
-              </button>
-            </div>
+              <NameWrapper>
+                <p>{photo?.name}</p>
+              </NameWrapper>
+              <SortButtonContainer>
+                <Button
+                  as="a"
+                  type="button"
+                  href={`/workmode/photo/${photo.id}`}
+                >
+                  Edit
+                </Button>
+                <Button
+                  type="button"
+                  data-idx={idx}
+                  onClick={handleRemovePhoto}
+                >
+                  Remove
+                </Button>
+              </SortButtonContainer>
+            </SortGridItem>
           </SortableItem>
         ))}
       </SortableList>
