+++ frontend/components/Error.js
@@ -1,4 +1,4 @@
 export default function Error({ content }) {
   if (!content) return null;
-  return <p>We are error!!</p>;
+  return <p>We got an error</p>;
 }
