+++ frontend/components/PleaseSignIn.js
@@ -1,0 +1,8 @@
+import { useUser } from '../hooks';
+import SignIn from './SignIn';
+
+export default function PleaseSignIn({ children }) {
+  const me = useUser();
+  if (!me) return <SignIn />;
+  return children;
+}
