+++ frontend/hooks/index.js
@@ -1,1 +1,7 @@
-//
+import { useQuery } from '@apollo/client';
+import { CURRENT_USER_QUERY } from '../graphql/queries';
+
+export function useUser() {
+  const { data } = useQuery(CURRENT_USER_QUERY);
+  return data?.authenticatedItem;
+}
