+++ frontend/pages/about.js
@@ -1,0 +1,9 @@
+import About from '../components/About';
+
+export default function AboutPage() {
+  return (
+    <>
+      <About />
+    </>
+  );
+}
