+++ frontend/pages/workmode/about.js
@@ -1,0 +1,11 @@
+import PleaseSignIn from '../../components/PleaseSignIn';
+
+export default function EditAboutPage() {
+  return (
+    <>
+      <PleaseSignIn>
+        <p>This is where you will edit the about page</p>
+      </PleaseSignIn>
+    </>
+  );
+}
