+++ frontend/pages/workmode/photo/add.js
@@ -1,9 +1,12 @@
 import CreatePhoto from '../../../components/CreatePhoto';
+import PleaseSignIn from '../../../components/PleaseSignIn';
 
 export default function AddPhotoPage() {
   return (
     <div>
-      <CreatePhoto />
+      <PleaseSignIn>
+        <CreatePhoto />
+      </PleaseSignIn>
     </div>
   );
 }
